styles refactor
