+++ src/components/CollapseIcon/CollapseIcon.js
@@ -23,7 +23,6 @@ const CollapseIcon = ({ isExpanded, className, ...restProps }) => {
       width="18"
       height="18"
       viewBox="0 0 18 18"
-      fill="none"
       {...restProps}
       className={mergedClassNames}
     >
