+++ src/components/CollapseIcon/index.js
@@ -1,0 +1,1 @@
+export { default as CollapseIcon } from './CollapseIcon';
