+++ src/components/UploadBar/style.scss
@@ -1,9 +1,11 @@
 @import "../../scss/variables/colors";
 @import "../../scss/variables/transitions";
+@import "../../scss/mixins/button";
 
-$base-class: "upload-bar";
+$upload-bar: "upload-bar";
+$file-upload-progress: "file-upload-progress";
 
-.#{$base-class} {
+.#{$upload-bar} {
   width: 100%;
   border-radius: 8px;
   box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.2);
@@ -22,24 +24,12 @@ $base-class: "upload-bar";
   }
 
   &--single {
-    .#{$base-class}__main {
+    .#{$upload-bar}__main {
       cursor: default;
     }
   }
 
-  &__success-icon {
-    margin-right: 8px;
-    display: flex;
-    align-items: center;
-    flex: 0 1 auto;
-
-    svg {
-      fill: #21a961;
-      width: 18px;
-      height: 18px;
-    }
-  }
-
+  &__success-icon,
   &__error-icon {
     margin-right: 8px;
     display: flex;
@@ -47,9 +37,20 @@ $base-class: "upload-bar";
     flex: 0 1 auto;
 
     svg {
-      fill: #d64646;
       width: 18px;
       height: 18px;
+    }
+  }
+
+  &__success-icon {
+    svg {
+      fill: #21a961;
+    }
+  }
+
+  &__error-icon {
+    svg {
+      fill: #d64646;
     }
   }
 
@@ -62,100 +63,6 @@ $base-class: "upload-bar";
 
     .progress-circle {
       margin-right: 8px;
-    }
-  }
-
-  &__actions {
-    margin-left: auto;
-    display: inline-flex;
-    align-items: center;
-    flex: 1 0 auto;
-    flex-wrap: nowrap;
-  }
-
-  &__close {
-    padding: 0;
-    margin: 0;
-    border: none;
-    margin-left: 8px;
-    display: flex;
-    justify-self: flex-end;
-    align-items: center;
-    flex: 0 1 auto;
-    color: $font-tertiary-color;
-    transition-property: opacity, border, color, background-color, box-shadow;
-    transition-duration: 200ms;
-    transition-timing-function: cubic-bezier(0.64, 0, 0.35, 1);
-    border-radius: 4px;
-
-    &:focus,
-    &:hover,
-    &:active {
-      outline: 0;
-    }
-
-    &:hover {
-      color: rgba(66, 77, 89, 1);
-    }
-
-    &:focus {
-      box-shadow: 0 0 1px 2px rgba($primary-color, 0.7);
-    }
-
-    &:active {
-      box-shadow: 0 0 0 1px $border-primary-color;
-      transition-duration: 100ms;
-    }
-
-    svg {
-      display: flex;
-      fill: currentColor;
-      height: 18px;
-      width: 18px;
-      cursor: pointer;
-    }
-  }
-
-  &__retry {
-    padding: 0;
-    margin: 0;
-    border: none;
-    margin-left: 8px;
-    display: flex;
-    align-items: center;
-    justify-self: flex-end;
-    flex: 0 1 auto;
-    color: $font-tertiary-color;
-    transition-property: opacity, border, color, background-color, box-shadow;
-    transition-duration: 200ms;
-    transition-timing-function: cubic-bezier(0.64, 0, 0.35, 1);
-    border-radius: 4px;
-
-    &:focus,
-    &:hover,
-    &:active {
-      outline: 0;
-    }
-
-    &:hover {
-      color: rgba(66, 77, 89, 1);
-    }
-
-    &:focus {
-      box-shadow: 0 0 1px 2px rgba($primary-color, 0.7);
-    }
-
-    &:active {
-      box-shadow: 0 0 0 1px $border-primary-color;
-      transition-duration: 100ms;
-    }
-
-    svg {
-      display: flex;
-      fill: currentColor;
-      height: 18px;
-      width: 18px;
-      cursor: pointer;
     }
   }
 
@@ -196,7 +103,7 @@ $base-class: "upload-bar";
     height: 100%;
     padding-right: 20px;
 
-    .file-upload-progress:not(:last-child) {
+    .#{$file-upload-progress}:not(:last-child) {
       margin-bottom: 14px;
       margin-top: 5px;
     }
@@ -212,50 +119,26 @@ $base-class: "upload-bar";
     flex: 0 1 auto;
     cursor: pointer;
     color: $font-secondary-color;
-    transition-property: opacity, border, color, background-color, box-shadow;
-    transition-duration: 200ms;
-    transition-timing-function: cubic-bezier(0.64, 0, 0.35, 1);
     border-radius: 4px;
 
-    &:focus,
-    &:hover,
-    &:active {
-      outline: 0;
-    }
-
-    &:hover {
-      color: rgba(66, 77, 89, 1);
-    }
-
-    &:focus {
-      box-shadow: 0 0 1px 2px rgba($primary-color, 0.7);
-    }
-
-    &:active {
-      box-shadow: 0 0 0 1px $border-primary-color;
-      transition-duration: 100ms;
-    }
+    @include button-transition;
+    @include icon-button-interactive-styles;
   }
 
   &--error {
-    .#{$base-class}__title {
+    .#{$upload-bar}__title {
       color: #d64646;
     }
   }
 }
 
-.file-upload-progress {
+.#{$file-upload-progress} {
   width: 100%;
   display: flex;
   align-items: center;
 
   &__actions {
-    margin-left: auto;
-    display: inline-flex;
     align-self: flex-start;
-    align-items: center;
-    flex: 1 0 auto;
-    flex-wrap: nowrap;
   }
 
   &__header {
@@ -289,23 +172,23 @@ $base-class: "upload-bar";
   }
 
   &--with-icon {
-    .file-upload-progress__wrapper {
+    .#{$file-upload-progress}__wrapper {
       width: calc(100% - 26px);
     }
   }
 
   &--error {
-    .file-upload-progress__title {
+    .#{$file-upload-progress}__title {
       color: #f4574c;
     }
   }
 
   &--success {
-    .file-upload-progress__title {
+    .#{$file-upload-progress}__title {
       margin-bottom: 0;
     }
 
-    .file-upload-progress__wrapper {
+    .#{$file-upload-progress}__wrapper {
       align-self: center;
     }
   }
@@ -322,83 +205,33 @@ $base-class: "upload-bar";
       height: 18px;
     }
   }
+}
 
-  &__close {
-    padding: 0;
-    margin: 0;
-    border: none;
-    margin-left: 8px;
-    display: flex;
-    justify-self: flex-end;
+.#{$file-upload-progress},
+.#{$upload-bar} {
+  &__actions {
+    margin-left: auto;
+    display: inline-flex;
     align-items: center;
-    flex: 0 1 auto;
-    color: $font-tertiary-color;
-    transition-property: opacity, border, color, background-color, box-shadow;
-    transition-duration: 200ms;
-    transition-timing-function: cubic-bezier(0.64, 0, 0.35, 1);
-    border-radius: 4px;
-
-    &:focus,
-    &:hover,
-    &:active {
-      outline: 0;
-    }
-
-    &:hover {
-      color: rgba(66, 77, 89, 1);
-    }
-
-    &:focus {
-      box-shadow: 0 0 1px 2px rgba($primary-color, 0.7);
-    }
-
-    &:active {
-      box-shadow: 0 0 0 1px $border-primary-color;
-      transition-duration: 100ms;
-    }
-
-    svg {
-      display: flex;
-      height: 18px;
-      width: 18px;
-      cursor: pointer;
-      fill: currentColor;
-    }
+    flex: 1 0 auto;
+    flex-wrap: nowrap;
   }
 
+  &__close,
   &__retry {
     padding: 0;
     margin: 0;
     border: none;
     margin-left: 8px;
     display: flex;
-    align-items: center;
     justify-self: flex-end;
+    align-items: center;
     flex: 0 1 auto;
     color: $font-tertiary-color;
-    transition-property: opacity, border, color, background-color, box-shadow;
-    transition-duration: 200ms;
-    transition-timing-function: cubic-bezier(0.64, 0, 0.35, 1);
     border-radius: 4px;
 
-    &:focus,
-    &:hover,
-    &:active {
-      outline: 0;
-    }
-
-    &:hover {
-      color: rgba(66, 77, 89, 1);
-    }
-
-    &:focus {
-      box-shadow: 0 0 1px 2px rgba($primary-color, 0.7);
-    }
-
-    &:active {
-      box-shadow: 0 0 0 1px $border-primary-color;
-      transition-duration: 100ms;
-    }
+    @include button-transition;
+    @include icon-button-interactive-styles;
 
     svg {
       display: flex;
@@ -406,37 +239,6 @@ $base-class: "upload-bar";
       height: 18px;
       width: 18px;
       cursor: pointer;
-    }
-  }
-}
-
-.collapse-icon {
-  display: inline-block;
-  flex-shrink: 0;
-  user-select: none;
-  fill: currentColor;
-
-  path {
-    transform-box: fill-box;
-    transform-origin: center;
-    transition: 0.3s cubic-bezier(0.175, 0.885, 0.32, 1.275) transform;
-  }
-
-  &__bottom-arrow {
-    transform: rotate(0);
-  }
-
-  &__top-arrow {
-    transform: rotate(0);
-  }
-
-  &--expanded {
-    .collapse-icon__bottom-arrow {
-      transform: rotate(-180deg);
-    }
-
-    .collapse-icon__top-arrow {
-      transform: rotate(180deg);
     }
   }
 }
